+++ Makefile.in
@@ -107,6 +107,7 @@ C_SRCS := $(srcroot)src/jemalloc.c \
 	$(srcroot)src/ehooks.c \
 	$(srcroot)src/eset.c \
 	$(srcroot)src/extent.c \
+	$(srcroot)src/extent2.c \
 	$(srcroot)src/extent_dss.c \
 	$(srcroot)src/extent_mmap.c \
 	$(srcroot)src/hash.c \
